Refactor Privileges to remove the need to pass session
