+++ src/components/authorization/policy/executor/edge-privileges.ts
@@ -1,7 +1,19 @@
-import { EnhancedResource, type ResourceShape, type Session } from '~/common';
+import { LazyGetter as Once } from 'lazy-get-decorator';
+import {
+  EnhancedResource,
+  type ResourceShape,
+  UnauthorizedException,
+} from '~/common';
 import { type ResourceObjectContext } from '../object.type';
-import { type PolicyExecutor } from './policy-executor';
-import { UserEdgePrivileges } from './user-edge-privileges';
+import {
+  type AllPermissionsOfEdgeView,
+  createAllPermissionsOfEdgeView,
+} from './all-permissions-view';
+import {
+  type FilterOptions,
+  type PolicyExecutor,
+  type ResolveParams,
+} from './policy-executor';
 
 export class EdgePrivileges<
   TResourceStatic extends ResourceShape<any>,
@@ -12,18 +24,105 @@ export class EdgePrivileges<
   constructor(
     resource: TResourceStatic | EnhancedResource<TResourceStatic>,
     readonly key: TKey,
+    private readonly object: ResourceObjectContext<TResourceStatic> | undefined,
     private readonly policyExecutor: PolicyExecutor,
   ) {
     this.resource = EnhancedResource.of(resource);
   }
 
-  forUser(session: Session, object?: ResourceObjectContext<TResourceStatic>) {
-    return new UserEdgePrivileges<TResourceStatic, TKey, TAction>(
+  /** @deprecated */
+  get session() {
+    return this.policyExecutor.sessionHost.current;
+  }
+
+  /** @deprecated Use {@link forContext} instead */
+  forUser(_session: unknown, object?: ResourceObjectContext<TResourceStatic>) {
+    return object ? this.forContext(object) : this;
+  }
+
+  get context() {
+    return this.object;
+  }
+
+  forContext(object: ResourceObjectContext<TResourceStatic>) {
+    if (object === this.object) {
+      return this;
+    }
+    return new EdgePrivileges(
       this.resource,
       this.key,
       object,
-      session,
       this.policyExecutor,
     );
   }
+
+  can(action: TAction) {
+    const perm = this.policyExecutor.resolve({
+      action,
+      resource: this.resource,
+      prop: this.key,
+    });
+    return perm === true || perm === false
+      ? perm
+      : perm.isAllowed({
+          object: this.object,
+          resource: this.resource,
+          session: this.policyExecutor.sessionHost.current,
+        });
+  }
+
+  verifyCan(action: TAction) {
+    if (this.can(action)) {
+      return;
+    }
+    throw UnauthorizedException.fromPrivileges(
+      action,
+      this.object,
+      this.resource,
+      this.key,
+    );
+  }
+
+  /**
+   * An alternative view that gives an object with all the permissions for
+   * actions.
+   * @example
+   * const privileges = Privileges.forEdge(User, 'email');
+   * if (privileges.all.read) {
+   *   // can read
+   * }
+   */
+  @Once()
+  get all(): AllPermissionsOfEdgeView<TAction> {
+    return createAllPermissionsOfEdgeView(this.resource, this);
+  }
+
+  /**
+   * Applies a filter to the `node` so that only readable nodes continue based on our polices.
+   * This requires `node` & `project` to be defined where this cypher snippet
+   * is inserted.
+   */
+  filterToReadable(options?: FilterOptions) {
+    return this.dbFilter({
+      action: 'read',
+      ...options,
+    });
+  }
+
+  dbFilter(options: FilterOptions & Pick<ResolveParams, 'action'>) {
+    return this.policyExecutor.cypherFilter({
+      ...options,
+      resource: this.resource,
+      prop: this.key,
+    });
+  }
 }
+
+/**
+ * @deprecated Use {@link EdgePrivileges} instead.
+ */
+export type UserEdgePrivileges<
+  TResourceStatic extends ResourceShape<any>,
+  TKey extends string,
+  TAction extends string,
+> = EdgePrivileges<TResourceStatic, TKey, TAction>;
