+++ src/components/authorization/policy/index.ts
@@ -11,9 +11,6 @@ export * from './builder/inherit-resource.helper';
 export * from './executor/privileges';
 export * from './executor/resource-privileges';
 export * from './executor/edge-privileges';
-export * from './executor/user-privileges';
-export * from './executor/user-resource-privileges';
-export * from './executor/user-edge-privileges';
 export type {
   AllPermissionsView,
   AllPermissionsOfEdgeView,
